requirements.txt
pyodbc>=4.0.39
python-decouple>=3.8
